% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAbnormal.R
\name{classifyAbnormal}
\alias{classifyAbnormal}
\title{Add classifications of potential abnormalities to input data.}
\usage{
classifyAbnormal(
  data,
  resolution_seconds = 60,
  N_days_required_daily_stats = 3,
  minimum_relval_per_hour = 0.1,
  epoch_size = 5,
  lowLuxThreshold = 50,
  maxLowLuxSequenceHours = 16
)
}
\arguments{
\item{data}{Tibble, that holds at least a Datetime and Lux column. \cr
Date time is expected to be a regular time series. \cr
If columns weight_X is included then this will be used as weighting for criteria X \cr
, e.g. weight_A for criteria A.}

\item{resolution_seconds}{Numeric, resolution in seconds at which to perform analysis. \cr
When epoch_size is small, consider setting this to a multitude of the \code{epoch_size} to speed up processing.}

\item{N_days_required_daily_stats}{Numeric, integer to indicate the minimum number \cr
of days required for deriving daily statistics}

\item{minimum_relval_per_hour}{Numeric, minimum relative variance per hour}

\item{epoch_size}{Numeric, epoch size  of input data in seconds}

\item{lowLuxThreshold}{Numeric, Lux value below which Lux is considered zero (darkness).}

\item{maxLowLuxSequenceHours}{Numeric, maximum number of hours with low lux}
}
\value{
Tibble data provided as input enhanced with classifications
}
\description{
Add classifications of potential abnormalities to input data.
}
