% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constantNonZeroLux.R
\name{constantNonZeroLux}
\alias{constantNonZeroLux}
\title{Classify periods with constant (low variantion) and non-zero Lux}
\usage{
constantNonZeroLux(x, minimum_relval_per_hour)
}
\arguments{
\item{x}{Tibble with columns min_per_hour, relvar_hour, and sd_diff_per_hour}

\item{minimum_relval_per_hour}{Numeric with the minimum relative variance in lux per hour}
}
\value{
Numeric vector with with length equal to x with classification  \cr
(1 = condition met, 0 = condition not met)
}
\description{
Classify periods with constant (low variantion) and non-zero Lux
}
