% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longLowValue.R
\name{longLowValue}
\alias{longLowValue}
\title{Classify long period with low value}
\usage{
longLowValue(x, x_threshold, window_size_hours, step_size, epoch_size)
}
\arguments{
\item{x}{Numeric vector with lux values}

\item{x_threshold}{Numeric, threshold to defined low}

\item{window_size_hours}{Numeric, rolling window size in hours as passed on to \link{rollApply}}

\item{step_size}{Numeric, integer as passed on to \link{rollApply}.}

\item{epoch_size}{Numeric, epoch size as passed on to \link{rollApply}.}
}
\value{
Numeric vector with with length equal to x with classification  \cr
(1 = condition met, 0 = condition not met)
}
\description{
Classify long period with low value
}
