% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbnormality.R
\name{plotAbnormality}
\alias{plotAbnormality}
\title{SPlot data with abnormality classifications}
\usage{
plotAbnormality(x)
}
\arguments{
\item{x}{Tibble, that holds at least columns Datetime, Lux, indicator_A, \cr
indicator_B, indicator_C, indicator_D, indicator_E, indicator_F}
}
\value{
A plot
}
\description{
Plot data with abnormality classifications
}
