% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollApply.R
\name{rollApply}
\alias{rollApply}
\title{Helper function to ease applying 1- and 16 hour rolling window functions}
\usage{
rollApply(data, window_size_hours, FUN, step_size, epoch_size)
}
\arguments{
\item{data}{Tibble, that holds at least a Datetime and Lux column}

\item{window_size_hours}{Numeric, rolling window size in hours}

\item{FUN}{Function to be applied per hour}

\item{step_size}{Numeric, integrer as used for  slider::slide_dbl argument .step.}

\item{epoch_size}{Numeric, epoch size  of input data in seconds}
}
\value{
Numeric vector with with length equal to the number of rows in data with function output
}
\description{
Helper function to ease applying 1- and 16 hour rolling window functions
}
