% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalStatistics.R
\name{temporalStatistics}
\alias{temporalStatistics}
\title{Derive temporal statistics from a tibble with time series}
\usage{
temporalStatistics(
  data,
  epoch_size = 5,
  lowLuxThreshold = 50,
  maxLowLuxSequenceHours = 16,
  step_size = NULL
)
}
\arguments{
\item{data}{Tibble, that holds at least a Datetime and Lux column}

\item{epoch_size}{Numeric, epoch size  of input data in seconds}

\item{lowLuxThreshold}{Numeric, Lux value below which Lux is considered zero (darkness).}

\item{maxLowLuxSequenceHours}{Numeric, maximum number of hours with low lux}

\item{step_size}{Numeric, used as input for function \link{rollApply}}
}
\value{
Tibble data provided as input enhanced with temporal statistics
}
\description{
Adds temporal statistics to tibble to a tibble with time series \cr
and derive a separate data.frame with daily statistics
}
