% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyClassifyAbnormal.R
\name{applyClassifyAbnormal}
\alias{applyClassifyAbnormal}
\title{Split data in segments if necessary and apply classifyAbnormal}
\usage{
applyClassifyAbnormal(
  data,
  resolution_seconds = 60,
  N_days_required_daily_stats = 3,
  minimum_relval_per_hour = 0.1,
  epoch_size = 5,
  lowLuxThreshold = 50,
  maxLowLuxSequenceHours = 16,
  plot_path = NULL,
  plot_id = "",
  userStatsFunction = NULL,
  userCriteriaFunction = NULL
)
}
\arguments{
\item{data}{Tibble, that holds at least a Datetime and Lux column. \cr
Date time is expected to be a regular time series. \cr
If columns weight_X is included then this will be used as weighting for criteria X \cr
, e.g. weight_A for criteria A.}

\item{resolution_seconds}{Numeric, resolution in seconds as passed on to \link{classifyAbnormal}}

\item{N_days_required_daily_stats}{Numeric, integer to indicate the minimum number \cr
of days required for deriving daily statistics as passed on to \link{classifyAbnormal}}

\item{minimum_relval_per_hour}{Numeric, minimum relative variance per hour as passed on to \link{classifyAbnormal}}

\item{epoch_size}{Numeric, epoch size  of input data in seconds}

\item{lowLuxThreshold}{Numeric, Lux value below which Lux is considered zero (darkness) as passed on to \link{classifyAbnormal}.}

\item{maxLowLuxSequenceHours}{Numeric, maximum number of hours with low lux as passed on to \link{classifyAbnormal}}

\item{plot_path}{Character, to indicate the folder path where to store output plot in png format}

\item{plot_id}{Character or numeric to be appended to the end of the plot file name}

\item{userStatsFunction}{function as provide by user for deriving temporal statistics}

\item{userCriteriaFunction}{function as provide by user for applying a new criteria}
}
\value{
Tibble data provided as input enhanced with classifications
}
\description{
Split data in segments if necessary and apply classifyAbnormal
}
